+++ src/App.js
@@ -3,7 +3,7 @@ import { HashRouter as Router, Route } from "react-router-dom";
 import "react-datepicker/dist/react-datepicker.css";
 import "./App.scss";
 import TransientTaxForm from "./components/forms/TransientTaxForm";
-import ConfirmationPage from "./components/forms/ConfirmationPage";
+import ConfirmationPage from "./pages/ConfirmationPage";
 import { ConstantsProvider } from "./context/ConstantsContext";
 import { Config } from "@baltimorecounty/javascript-utilities";
 const { setConfig } = Config;
